Revert "Merge pull request #88 from yargevad/ISSUE-87"
This reverts commit 3b11e039dc802bfc865cd2e515583e9b0911aaf0, reversing
changes made to 586415a7db9d2b1538cd2c05ca2dbbce0ee9cc62.

This merge actually caused the tests to fail. After digging in I decided
that the documentation is good enough as a test on creation of Authboss
is not possible with the current API. And this being only in the
register module was kind of a code smell.