Fix bug in expire
Users were able to persist one request past expiration because the
session state was not actually hidden for downstream http handlers as it
was in Authboss 1.0 because of the abstraction of ClientState where an
http handler cannot alter the session subsequent handlers.

There's a number of positive reasons for the design as is, but it did
end up creating this bug when this code was ported from v1 to v2.